requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-graphql-jwt==0.3.4
django-jazzmin==2.5.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
mysqlclient==2.1.1
promise==2.3
pycodestyle==2.9.1
PyJWT==2.5.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tzdata==2022.4
